requirements.txt
Pillow>=9.0
requests>=2.28
